templates/python/custom-agent/pyproject.toml
[tool.poetry]
name = "custom-agent"
version = "1.0.0"
description = "Template pour agent personnalisé avec BaseAgent Google ADK"
authors = [{name = "Google ADK Community"}]
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry.dependencies]
python = ">=3.11"
google-adk = ">=1.3.0"
python-dotenv = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
